# Added Buffer & Array equality check and added includeDiffOnDataEquali…