… test/fixtures/5a.png
… test/fixtures/5diff.png